requirements.txt
black==24.2.0
click==8.1.7
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pytest==8.0.2
PyYAML==6.0.1
